pyproject.toml
[tool.poetry]
name = "resume-backend"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">3.10,<3.12"
Django = "^4.2"
django-configurations = "^2.4"
dj-database-url = "^2.1.0"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
django-extensions = "^3.2.1"
django-import-export = "^3.3.1"
django-storages = "^1.14.1"
Pillow = "^10.0.1"
django-cors-headers = "^4.2.0"
django-querycount = "^0.8.3"
gunicorn = "^21.2.0"
djangorestframework-simplejwt = "^5.2.1"
psycopg = "^3.1.12"
drf-nested-routers = "^0.93.4"
drf-spectacular = "^0.26.5"
boto3 = ">=1.28.60"

#whitenoise = ">=6.2.0" #uncomment in heroku
djangorestframework-camel-case = "^1.4.2"
django-tinymce = "^3.6.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
ruff = "^0.0.292"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

# To never enforce `E501` (line length violations) uncomment the next line
# ignore = ["E501"]
